added spawn timers to mcp increased speed of potholes changed nitroSpd from int to float

+++ Pothole Dodger v.2/Assets/Scripts/Config.cs
@@ -4,14 +4,14 @@ using System.Collections;
 public class Config
 {
 	private static float mapMovSpd = 1f;
-	private static int nitroSpd = 2;
+	private static float nitroSpd = 2f;
 
 	public static float GetMapMovSpd()
 	{
 		return mapMovSpd;
 	}
 	
-	public static int GetNitroSpd()
+	public static float GetNitroSpd()
 	{
 		return nitroSpd;
 	}
